+++ packages/license-server/src/index.js
@@ -1,2 +1,3 @@
 export { createInMemoryLeaseStore } from "./storage.js";
+export { createSqliteLeaseStore } from "./storage-sqlite.js";
 export { createLicenseFetchHandler, startLicenseServer } from "./server.js";